+++ lib/store/store_test.go
@@ -1,8 +1,9 @@
-package libunlynx_test
+package store_test
 
 import (
 	"github.com/dedis/kyber/util/random"
 	"github.com/lca1/unlynx/lib"
+	"github.com/lca1/unlynx/lib/store"
 	"github.com/stretchr/testify/assert"
 	"strconv"
 	"testing"
@@ -55,7 +56,7 @@ func TestStoring(t *testing.T) {
 	where := []libunlynx.WhereQueryAttribute{{Name: "0", Value: libunlynx.CipherText{}}, {Name: "1", Value: libunlynx.CipherText{}}}
 
 	// Constructor Test
-	storage := libunlynx.NewStore()
+	storage := store.NewStore()
 
 	// (1) Test Insert and Pull DpResponses
 	storage.InsertDpResponse(libunlynx.DpResponse{GroupByEnc: testEncMap, WhereClear: testClearMap, AggregatingAttributesEnc: testAggrMap1}, true, groupBy, sum, where)
@@ -81,7 +82,7 @@ func TestStoring(t *testing.T) {
 	storage.PullLocallyAggregatedResponses()
 
 	// (4) Test Insert and Pull DpResponses but with different parameters
-	storage = libunlynx.NewStore()
+	storage = store.NewStore()
 
 	storage.InsertDpResponse(libunlynx.DpResponse{GroupByClear: testClearMap, GroupByEnc: testEncMap, WhereClear: testClearMap, WhereEnc: testEncMap, AggregatingAttributesEnc: testAggrMap2}, true, groupBy, sum, where)
 	storage.InsertDpResponse(libunlynx.DpResponse{GroupByEnc: testEncMap, AggregatingAttributesEnc: testAggrMap2}, false, groupBy, sum, where)
